requirements.txt
fastapi[standard]
uvicorn
pydantic
azure-identity
psycopg2
passlib
langchain[openai]
langgraph
langchain-community 
beautifulsoup4